+++ react/index.tsx
@@ -1,121 +1,24 @@
 /* global module */
 import 'core-js/es6/symbol'
 import 'core-js/fn/symbol/iterator'
-import { prop } from 'ramda'
 import { canUseDOM } from 'exenv'
 import * as runtimeGlobals from './core/main'
-import { createReactIntl } from './utils/reactIntl'
 
-import { createCustomReactApollo } from './utils/reactApollo'
-import { fetchUncriticalStyles, UncriticalStyle } from './utils/assets'
+import { polyfillIntl } from './start/intl'
+import { addAMPProxy } from './start/amp'
+import { patchLibs } from './start/patchLibs'
+import { registerRuntimeGlobals } from './start/register'
+import { createUncriticalPromise } from './start/uncritical'
+import { loadRuntimeJSONs } from './start/runtime'
 
-function initJSON() {
-  const scripts = window?.document?.querySelectorAll<HTMLTemplateElement>(
-    'template[data-type="json"]'
-  )
-  if (!scripts || scripts.length === 0) {
-    return Promise.resolve()
-  }
+const intlPolyfillPromise = polyfillIntl()
+registerRuntimeGlobals(runtimeGlobals)
+patchLibs()
 
-  const promises = Array.from(scripts).map(
-    (script) =>
-      new Promise((resolve) => {
-        setTimeout(() => {
-          let value = ''
-          const childNodes = script.content.childNodes
-          for (let i = 0; i < childNodes.length; i++) {
-            const node = childNodes[i]
-            value += node.nodeValue
-          }
-          setTimeout(() => {
-            ;(window as any)[script.id] = JSON.parse(value)
-            resolve()
-          }, 1)
-        }, 1)
-      })
-  )
-
-  return new Promise((resolve) => {
-    Promise.all(promises).then(() => {
-      window.__RUNTIME__.extensions =
-        window.__RUNTIME_EXTENSIONS__ ?? window.__RUNTIME__.extensions
-      resolve()
-    })
-  })
-}
-
-let intlPolyfillPromise: Promise<void> = Promise.resolve()
-
-function init() {
-  window.__RENDER_8_RUNTIME__ = { ...runtimeGlobals }
-
-  // compatibility
-  window.__RENDER_8_COMPONENTS__ =
-    window.__RENDER_8_COMPONENTS__ || global.__RENDER_8_COMPONENTS__
-  window.__RENDER_8_HOT__ = window.__RENDER_8_HOT__ || global.__RENDER_8_HOT__
-  global.__RUNTIME__ = window.__RUNTIME__
-
-  if (window.IntlPolyfill) {
-    window.IntlPolyfill.__disableRegExpRestore()
-    if (!window.Intl) {
-      window.Intl = window.IntlPolyfill
-    }
-  }
-  if (
-    window.Intl &&
-    canUseDOM &&
-    (!window.Intl.PluralRules || !window.Intl.RelativeTimeFormat)
-  ) {
-    intlPolyfillPromise = import('./intl-polyfill').then(prop('default'))
-  }
-
-  if (module.hot) {
-    module.hot.accept('./core/main', () => {
-      // eslint-disable-next-line @typescript-eslint/no-var-requires
-      const hotGlobals = require('./core/main')
-      window.__RENDER_8_RUNTIME__.ExtensionContainer =
-        hotGlobals.ExtensionContainer
-      window.__RENDER_8_RUNTIME__.ExtensionPoint = hotGlobals.ExtensionPoint
-      window.__RENDER_8_RUNTIME__.LayoutContainer = hotGlobals.LayoutContainer
-      window.__RENDER_8_RUNTIME__.Link = hotGlobals.Link
-      window.__RENDER_8_RUNTIME__.Loading = hotGlobals.Loading
-      window.__RENDER_8_RUNTIME__.buildCacheLocator =
-        hotGlobals.buildCacheLocator
-      runtimeGlobals.start()
-    })
-  }
-
-  if (!window.__RUNTIME__.amp) {
-    window.ReactAMPHTML = window.ReactAMPHTMLHelpers =
-      typeof Proxy !== 'undefined'
-        ? new Proxy(
-            {},
-            {
-              get: (_, key) => {
-                if (key === '__esModule' || key === 'constructor') {
-                  return
-                }
-
-                const message = canUseDOM
-                  ? 'You can not render AMP components on client-side'
-                  : 'You must check runtime.amp to render AMP components'
-
-                throw new Error(message)
-              },
-            }
-          )
-        : {} // IE11 users will not have a clear error in this case
-  }
-
-  if (window.ReactApollo) {
-    window.ReactApollo = createCustomReactApollo()
-  }
-
-  if (window.ReactIntl) {
-    window.ReactIntl = createReactIntl()
-  }
-}
 function start() {
+  global.__RUNTIME__ = window.__RUNTIME__
+  addAMPProxy(window.__RUNTIME__)
+
   if (window.__RUNTIME__.start && !window.__ERROR__) {
     if (canUseDOM) {
       const contentLoadedPromise =
@@ -124,7 +27,9 @@ function start() {
           window.addEventListener('DOMContentLoaded', resolve)
         )
 
-      const resolveUncriticalPromise = createUncriticalPromise()
+      const resolveUncriticalPromise = createUncriticalPromise(
+        window.__RUNTIME__
+      )
       Promise.all([contentLoadedPromise, intlPolyfillPromise]).then(() => {
         setTimeout(() => {
           window?.performance?.mark?.('render-start')
@@ -144,102 +49,23 @@ function start() {
   }
 }
 
-function createUncriticalPromise() {
-  const {
-    __RUNTIME__: { uncriticalStyleRefs },
-  } = window
-  const criticalElement = document.querySelector('style#critical')
-  let resolve = () => {}
-
-  if (!uncriticalStyleRefs || !criticalElement) {
-    return resolve
-  }
-
-  window.__UNCRITICAL_PROMISE__ = new Promise<void>((r) => {
-    resolve = r
+if (module.hot) {
+  module.hot.accept('./core/main', () => {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const hotGlobals = require('./core/main')
+    window.__RENDER_8_RUNTIME__.ExtensionContainer =
+      hotGlobals.ExtensionContainer
+    window.__RENDER_8_RUNTIME__.ExtensionPoint = hotGlobals.ExtensionPoint
+    window.__RENDER_8_RUNTIME__.LayoutContainer = hotGlobals.LayoutContainer
+    window.__RENDER_8_RUNTIME__.Link = hotGlobals.Link
+    window.__RENDER_8_RUNTIME__.Loading = hotGlobals.Loading
+    window.__RENDER_8_RUNTIME__.buildCacheLocator = hotGlobals.buildCacheLocator
+    runtimeGlobals.start()
   })
-    .then(() => {
-      const { base = [], overrides = [] } = uncriticalStyleRefs
-      return fetchUncriticalStyles([...base, ...overrides])
-    })
-    .then((uncriticalStyles) => {
-      if (!uncriticalStyles) {
-        console.error('Missing lazy links')
-        return
-      }
-
-      const debugCriticalCSS = window.__RUNTIME__.query?.__debugCriticalCSS
-
-      const createUncriticalStyle = (uncriticalStyle: UncriticalStyle) => {
-        if (!uncriticalStyle) {
-          return
-        }
-        const style = document.createElement('style')
-
-        style.id = uncriticalStyle.id ?? ''
-        style.className = `uncritical ${uncriticalStyle.className ?? ''}`
-        style.media = uncriticalStyle.media
-        style.innerHTML = uncriticalStyle.body
-        style.setAttribute('data-href', uncriticalStyle.href)
-
-        document.head.appendChild(style)
-      }
-
-      const clearCritical = () => {
-        if (criticalElement.parentElement) {
-          criticalElement.remove()
-        }
-      }
-
-      const applyUncritical = () => {
-        uncriticalStyles.forEach(createUncriticalStyle)
-        clearCritical()
-      }
-
-      /** Doesn't apply uncritical CSS automatically--exposes functions
-       * to the window to manually do it, for debugging purposes
-       */
-      if (debugCriticalCSS === 'manual') {
-        ;(window as any).applyUncritical = applyUncritical
-        ;(window as any).clearCritical = clearCritical
-
-        let currentUncritical = 0
-        ;(window as any).stepUncritical = () => {
-          if (currentUncritical === -1) {
-            console.log('Uncritical has finished being applied.')
-          }
-          const current = uncriticalStyles[currentUncritical]
-          if (!current) {
-            console.log(
-              'All uncritical styles applied. Cleaning critical styles.'
-            )
-            clearCritical()
-            currentUncritical = -1
-          }
-          console.log('Applying uncritical style', current)
-          createUncriticalStyle(current)
-          currentUncritical++
-        }
-
-        console.log(
-          `Run the following functions on the console to manually apply uncritical CSS:
-            - applyUncritical()
-            - stepUncritical()
-            - clearCritical()
-          `
-        )
-      } else {
-        applyUncritical()
-      }
-    })
-
-  return resolve
 }
-
-init()
 
 if (!canUseDOM) {
   start()
 } else {
-  initJSON().then(() => start())
+  loadRuntimeJSONs().then(() => start())
 }
